Rename mal to myanimelist_extractor to mitigate problems with windows defender
