+++ go.mod
@@ -1,3 +1,3 @@
-module sxeix/mal
+module sxeix/myanimelist_extractor
 
 go 1.16
